clean catalog codes (#15938)
clean codes

Approved by: @XuPeng-SH
